mxshop_srvs/user_srv/requirements.txt
peewee
pymysql
passlib
grpcio==1.32.0
grpcio-tools==1.32.0
protobuf==3.13.0
